+++ lib/xmat-action-text/ts/xmat-action-text.directive.ts
@@ -10,10 +10,12 @@ const DEFAULT_COLOR = "primary";
 // Boilerplate for applying mixins to MatToolbar.
 /** @docs-private */
 export class XmatActionTextBase {
+    // tslint:disable-next-line:naming-convention
     constructor(public _elementRef: ElementRef) {
     }
 }
 
+// tslint:disable-next-line:naming-convention
 export const _XmatActionMixinBase: Constructor<CanColor> & typeof XmatActionTextBase = mixinColor(XmatActionTextBase);
 
 @Directive({
@@ -32,6 +34,7 @@ export class XmatActionTextDirective extends _XmatActionMixinBase implements Can
 
     @Input() disabled: boolean;
 
+    // tslint:disable-next-line:naming-convention
     constructor(public _elementRef: ElementRef) {
         super(_elementRef);
     }